+++ packages/did-ui-react/src/components/Ramp/components/RampPreview/component/RampSell/index.tsx
@@ -69,7 +69,39 @@ export default function RampSell({
     icon: routerIcon,
     address: routerAddress,
   } = selectedCrypto;
+  const [init, setInit] = useState(false);
+  const containerRef = useRef<HTMLDivElement>(null);
+  const fiatTextRef = useRef<HTMLDivElement>(null);
+  const [maxWidth, setMaxWidth] = useState(0);
 
+  useEffect(() => {
+    if (!init) {
+      return;
+    }
+    const res = setTimeout(() => {
+      if (containerRef.current && fiatTextRef.current) {
+        const style = window.getComputedStyle(containerRef.current);
+        const fiatTextStyle = window.getComputedStyle(fiatTextRef.current);
+        const paddingLeft = parseFloat(style.paddingLeft || '0');
+        const paddingRight = parseFloat(style.paddingRight || '0');
+        const marginLeft = parseFloat(fiatTextStyle.marginLeft || '0');
+        const paddingWidth = paddingLeft + paddingRight;
+        console.log(
+          containerRef.current.clientWidth,
+          paddingLeft,
+          paddingRight,
+          fiatTextRef.current.clientWidth,
+          marginLeft,
+        );
+        const noPaddingWidth =
+          containerRef.current.clientWidth - paddingWidth - fiatTextRef.current.clientWidth - marginLeft;
+        setMaxWidth(noPaddingWidth);
+      }
+    }, 100);
+    return () => {
+      clearTimeout(res);
+    };
+  }, [init]);
   const textInputRef = useRef<HTMLInputElement>(null);
   const [fiatList, setFiatList] = useState<IRampFiatItem[]>([]);
   const [buttonLoading, setButtonLoading] = useState(false);
@@ -346,6 +378,7 @@ export default function RampSell({
       console.log('sellForm refreshList error', error);
     } finally {
       setLoading(false);
+      setInit(true);
       if (textInputRef.current) {
         textInputRef.current.focus();
       }
@@ -482,7 +515,7 @@ export default function RampSell({
         }
         rightCallback={onChangeCurrency}
       />
-      <div className="portkey-ui-ramp-content portkey-ui-flex-column-center ramp-sell">
+      <div className="portkey-ui-ramp-content portkey-ui-flex-column-center ramp-sell" ref={containerRef}>
         <div className="ramp-sell-pageWrap">
           <MaxTokenShortcut tokenInfo={currency.crypto} maxAmount={maxAmount} onMaxPress={onMaxPress} />
           {/* Fiat Input Section */}
@@ -494,10 +527,12 @@ export default function RampSell({
               amount={amount}
               amountError={amountError}
               textInputRef={textInputRef}
+              maxWidth={maxWidth}
             />
             {/* Fiat Symbol */}
             <div
               className="fiatText"
+              ref={fiatTextRef}
               onClick={() => {
                 if (textInputRef.current) {
                   textInputRef.current.focus();
@@ -542,7 +577,6 @@ export default function RampSell({
       <CommonModal
         className="change-currency-modal-list"
         open={openFiatModal}
-        height={582}
         onClose={() => {
           setOpenFiatModal(false);
         }}>
@@ -557,7 +591,7 @@ export default function RampSell({
             }}
           />
         </div>
-        <ul style={{ width: '100%', marginTop: 16, overflowY: 'auto' }}>
+        <ul style={{ width: '100%', overflowY: 'auto' }}>
           {filteredList?.map((item, index) => (
             <li
               key={index + '_' + item.symbol}