+++ packages/did-ui-react/src/components/SignStep/components/Step1/index.tsx
@@ -51,7 +51,7 @@ function Step1({
   const onConfirm = useCallback(() => {
     if (!signInSuccessRef.current) return setOpen(false);
     const createType = signInSuccessRef.current.isLoginGuardian ? 'Login' : 'SignUp';
-    setCreateType(createType);
+    // setCreateType(createType);
     onSignInFinished?.({
       isFinished: false,
       result: {
@@ -65,33 +65,44 @@ function Step1({
   const onSuccess = useCallback(
     async (value: IGuardianIdentifierInfo) => {
       signInSuccessRef.current = value;
-      if (!value.isLoginGuardian) {
-        setLoading(false);
-        const isContinue = await onSignUpHandlerRef.current?.({
-          identifier: value.identifier,
-          accountType: value.accountType,
-          authenticationInfo: {
-            authToken: value.authenticationInfo?.authToken,
-            idToken: value.authenticationInfo?.idToken,
-            nonce: value.authenticationInfo?.nonce,
-            timestamp: value.authenticationInfo?.timestamp,
+      try {
+        // setLoading(true);
+        if (!value.isLoginGuardian) {
+          // setLoading(false)
+          const isContinue = await onSignUpHandlerRef.current?.({
+            identifier: value.identifier,
+            accountType: value.accountType,
+            authenticationInfo: {
+              authToken: value.authenticationInfo?.authToken,
+              idToken: value.authenticationInfo?.idToken,
+              nonce: value.authenticationInfo?.nonce,
+              timestamp: value.authenticationInfo?.timestamp,
+            },
+          });
+          if (isContinue === SignUpValue.otherSeverRegisterButContinue) {
+            return onConfirm();
+          }
+          if (isContinue === SignUpValue.cancelRegister) {
+            return;
+          }
+          if (createType !== 'SignUp') {
+            return setOpen(true);
+          }
+        }
+
+        if (value.isLoginGuardian && createType !== 'Login') return setOpen(true);
+
+        await onSignInFinished?.({
+          isFinished: false,
+          result: {
+            type: createType,
+            value,
           },
         });
-        if (isContinue === SignUpValue.otherSeverRegisterButContinue) return onConfirm();
-        if (isContinue === SignUpValue.cancelRegister) return;
-        if (createType !== 'SignUp') return setOpen(true);
+        setOpen(false);
+      } finally {
+        // setLoading(false);
       }
-
-      if (value.isLoginGuardian && createType !== 'Login') return setOpen(true);
-
-      onSignInFinished?.({
-        isFinished: false,
-        result: {
-          type: createType,
-          value,
-        },
-      });
-      setOpen(false);
     },
     [createType, onConfirm, onSignInFinished],
   );
@@ -124,7 +135,6 @@ function Step1({
   //   // Get phoneCountry by service, update phoneCountry
   //   getPhoneCountry();
   // }, [getPhoneCountry]);
-
   return (
     <>
       {design === Design.SocialDesign && (
@@ -136,6 +146,7 @@ function Step1({
           isErrorTip={isErrorTip}
           onError={onError}
           onSuccess={onSuccess}
+          onSignTypeChange={setCreateType}
         />
       )}