+++ packages/did-ui-react/src/components/context/index.tsx
@@ -8,7 +8,6 @@ import { useEffectOnce } from 'react-use';
 import { did } from '../../utils';
 import ConfigProvider from '../config-provider';
 import { initConfig } from './initConfig';
-import { initTheme } from '../../assets/theme';
 
 const INITIAL_STATE = {
   theme: 'light',
@@ -17,7 +16,8 @@ const INITIAL_STATE = {
 const PortkeyContext = createContext<any>(INITIAL_STATE);
 
 export function usePortkey(): [PortkeyState, BasicActions] {
-  return useContext(PortkeyContext);
+  const context = useContext(PortkeyContext);
+  return context;
 }
 
 //reducer
@@ -44,6 +44,7 @@ export interface ProviderProps {
 }
 export default function Provider({ theme, chainType, sandboxId, networkType, children }: ProviderProps) {
   const [state, dispatch] = useReducer(reducer, INITIAL_STATE);
+  console.log('setGlobalConfig BaseConfigProvider=>Provider theme', theme);
   useEffectOnce(() => {
     initConfig();
     if (did.config.storageMethod) {
@@ -55,8 +56,8 @@ export default function Provider({ theme, chainType, sandboxId, networkType, chi
     ConfigProvider.setGlobalConfig({ networkType });
   }, [networkType]);
   useEffect(() => {
-    ConfigProvider.setGlobalConfig({ theme });
-    initTheme(theme);
+    console.log('setGlobalConfig1', theme);
+    theme && ConfigProvider.setTheme(theme);
   }, [theme]);
 
   return (